requirements.txt
astroid==2.1.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.2
django-oauth-toolkit==1.2.0
djangorestframework==3.9.1
entrypoints==0.3
flake8==3.7.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mypy==0.660
mypy-extensions==0.4.1
oauthlib==3.0.1
psycopg2-binary==2.8.1
pycodestyle==2.5.0
pyflakes==2.1.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.9
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1
